Move filter table to syncapi where it is used
